+++ packages/browser/src/index.ts
@@ -142,7 +142,10 @@ export default function Browser(options: Options = {}): History {
         fn();
       });
     },
-    update: function update(to: ToArgument, navType: NavType = "ANCHOR"): void {
+    navigate: function navigate(
+      to: ToArgument,
+      navType: NavType = "ANCHOR"
+    ): void {
       let setup: NavSetup;
       const location = createLocation(to);
       switch (navType) {