+++ packages/browser/tests/unit/anchor.spec.ts
@@ -15,7 +15,7 @@ function ignoreFirstCall(fn) {
 
 // We create our own jsdom instead of using the one that Jest will create
 // so that we can reset the DOM between tests
-describe('update(..., "ANCHOR")', () => {
+describe('navigate(..., "ANCHOR")', () => {
   let dom;
   let window;
 
@@ -35,7 +35,7 @@ describe('update(..., "ANCHOR")', () => {
   describe("without a response handler", () => {
     it("does nothing", () => {
       const testHistory = Browser();
-      testHistory.update("/two");
+      testHistory.navigate("/two");
       expect(window.location.pathname).toBe("/one");
     });
   });
@@ -45,7 +45,7 @@ describe('update(..., "ANCHOR")', () => {
       const testHistory = Browser();
       const router = jest.fn();
       testHistory.respondWith(router); // calls router
-      testHistory.update("/two");
+      testHistory.navigate("/two");
       expect(router.mock.calls.length).toBe(2);
     });
 
@@ -59,7 +59,7 @@ describe('update(..., "ANCHOR")', () => {
           });
         });
         testHistory.respondWith(router);
-        testHistory.update("/two?test=ing");
+        testHistory.navigate("/two?test=ing");
       });
 
       it("is a location object created from pushed object", () => {
@@ -71,7 +71,7 @@ describe('update(..., "ANCHOR")', () => {
           });
         });
         testHistory.respondWith(router);
-        testHistory.update({ pathname: "/two", hash: "test" });
+        testHistory.navigate({ pathname: "/two", hash: "test" });
       });
 
       it("(new) increments major key", () => {
@@ -84,7 +84,7 @@ describe('update(..., "ANCHOR")', () => {
         const [initMajor] = testHistory.location.key.split(".");
         const initMajorNum = parseInt(initMajor, 10);
 
-        testHistory.update("/next");
+        testHistory.navigate("/next");
 
         const current = testHistory.location;
         const [currentMajor, currentMinor] = current.key.split(".");
@@ -105,7 +105,7 @@ describe('update(..., "ANCHOR")', () => {
         const initMajorNum = parseInt(initMajor, 10);
         const initMinorNum = parseInt(initMinor, 10);
 
-        testHistory.update("/one");
+        testHistory.navigate("/one");
 
         const current = testHistory.location;
         const [currentMajor, currentMinor] = current.key.split(".");
@@ -125,7 +125,7 @@ describe('update(..., "ANCHOR")', () => {
           pending.finish();
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/next");
+        testHistory.navigate("/next");
       });
 
       it('(same) is "REPLACE"', () => {
@@ -135,7 +135,7 @@ describe('update(..., "ANCHOR")', () => {
           pending.finish();
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/one");
+        testHistory.navigate("/one");
       });
     });
 
@@ -148,15 +148,15 @@ describe('update(..., "ANCHOR")', () => {
           expect(window.location.pathname).toBe("/two");
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two");
+        testHistory.navigate("/two");
       });
 
-      it("does nothing if update is not called", () => {
+      it("does nothing if navigate() is not called", () => {
         const testHistory = Browser();
         let call = 0;
         function router(pending) {}
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two");
+        testHistory.navigate("/two");
         expect(window.location.pathname).toBe("/one");
       });
 
@@ -166,7 +166,7 @@ describe('update(..., "ANCHOR")', () => {
           pending.finish();
         }
         testHistory.respondWith(router); // calls router
-        testHistory.update("/next");
+        testHistory.navigate("/next");
         expect(testHistory.location).toMatchObject({
           pathname: "/next"
         });
@@ -178,7 +178,7 @@ describe('update(..., "ANCHOR")', () => {
           pending.finish();
         }
         testHistory.respondWith(router); // calls router
-        testHistory.update("/next");
+        testHistory.navigate("/next");
         expect(testHistory.action).toBe("PUSH");
       });
 
@@ -188,7 +188,7 @@ describe('update(..., "ANCHOR")', () => {
           pending.finish();
         }
         testHistory.respondWith(router); // calls router
-        testHistory.update("/one");
+        testHistory.navigate("/one");
         expect(testHistory.action).toBe("REPLACE");
       });
     });
@@ -202,7 +202,7 @@ describe('update(..., "ANCHOR")', () => {
           expect(window.location.pathname).toBe("/one");
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two");
+        testHistory.navigate("/two");
       });
 
       it("does not update the action value", () => {
@@ -213,7 +213,7 @@ describe('update(..., "ANCHOR")', () => {
           expect(testHistory.action).toBe("PUSH");
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two");
+        testHistory.navigate("/two");
       });
     });
   });
@@ -228,7 +228,7 @@ describe('update(..., "ANCHOR")', () => {
       testHistory.confirmWith(confirm);
       testHistory.respondWith(router); // calls router
 
-      testHistory.update("/next");
+      testHistory.navigate("/next");
       expect(router.mock.calls.length).toBe(2);
     });
 
@@ -241,7 +241,7 @@ describe('update(..., "ANCHOR")', () => {
       testHistory.confirmWith(confirm);
       testHistory.respondWith(router); // calls router
 
-      testHistory.update("/next");
+      testHistory.navigate("/next");
       expect(router.mock.calls.length).toBe(1);
     });
   });