+++ packages/hash/tests/unit/push.spec.ts
@@ -15,7 +15,7 @@ function ignoreFirstCall(fn) {
 
 // We create our own jsdom instead of using the one that Jest will create
 // so that we can reset the DOM between tests
-describe('update(..., "PUSH")', () => {
+describe('navigate(..., "PUSH")', () => {
   let dom;
   let window;
 
@@ -35,7 +35,7 @@ describe('update(..., "PUSH")', () => {
   describe("without a response handler", () => {
     it("does nothing", () => {
       const testHistory = Hash();
-      testHistory.update("/two", "PUSH");
+      testHistory.navigate("/two", "PUSH");
       expect(window.location.hash).toBe("#/one");
     });
   });
@@ -45,7 +45,7 @@ describe('update(..., "PUSH")', () => {
       const testHistory = Hash();
       const router = jest.fn();
       testHistory.respondWith(router); // calls router
-      testHistory.update("/two", "PUSH");
+      testHistory.navigate("/two", "PUSH");
       expect(router.mock.calls.length).toBe(2);
     });
 
@@ -59,7 +59,7 @@ describe('update(..., "PUSH")', () => {
           });
         });
         testHistory.respondWith(router);
-        testHistory.update("/two?test=ing", "PUSH");
+        testHistory.navigate("/two?test=ing", "PUSH");
       });
 
       it("is a location object created from pushed object", () => {
@@ -71,7 +71,7 @@ describe('update(..., "PUSH")', () => {
           });
         });
         testHistory.respondWith(router);
-        testHistory.update({ pathname: "/two", hash: "test" }, "PUSH");
+        testHistory.navigate({ pathname: "/two", hash: "test" }, "PUSH");
       });
 
       it("increments current major key value by 1, sets minor value to 0", () => {
@@ -84,7 +84,7 @@ describe('update(..., "PUSH")', () => {
         const [initMajor] = testHistory.location.key.split(".");
         const initMajorNum = parseInt(initMajor, 10);
 
-        testHistory.update("/next", "PUSH");
+        testHistory.navigate("/next", "PUSH");
 
         const current = testHistory.location;
         const [currentMajor, currentMinor] = current.key.split(".");
@@ -104,7 +104,7 @@ describe('update(..., "PUSH")', () => {
             pending.action === "POP" &&
             pending.location.pathname === "/three"
           ) {
-            testHistory.update("/new-four", "PUSH");
+            testHistory.navigate("/new-four", "PUSH");
             afterPop = true;
           }
           if (afterPop) {
@@ -114,10 +114,10 @@ describe('update(..., "PUSH")', () => {
         }
         testHistory.respondWith(router); // calls router
         // just getting the history in the expected state
-        testHistory.update("/two", "PUSH"); // 1.0
-        testHistory.update("/three", "PUSH"); // 2.0
-        testHistory.update("/four", "PUSH"); // 3.0
-        testHistory.update("/five", "PUSH"); // 4.0
+        testHistory.navigate("/two", "PUSH"); // 1.0
+        testHistory.navigate("/three", "PUSH"); // 2.0
+        testHistory.navigate("/four", "PUSH"); // 3.0
+        testHistory.navigate("/five", "PUSH"); // 4.0
         testHistory.go(-2); // 2.0
       });
     });
@@ -130,7 +130,7 @@ describe('update(..., "PUSH")', () => {
           pending.finish();
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/next", "PUSH");
+        testHistory.navigate("/next", "PUSH");
       });
     });
 
@@ -143,15 +143,15 @@ describe('update(..., "PUSH")', () => {
           expect(window.location.hash).toBe("#/two");
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two", "PUSH");
+        testHistory.navigate("/two", "PUSH");
       });
 
-      it("does nothing if update is not called", () => {
+      it("does nothing if navigate is not called", () => {
         const testHistory = Hash();
         let call = 0;
         function router(pending) {}
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two", "PUSH");
+        testHistory.navigate("/two", "PUSH");
         expect(window.location.hash).toBe("#/one");
       });
 
@@ -161,7 +161,7 @@ describe('update(..., "PUSH")', () => {
           pending.finish();
         }
         testHistory.respondWith(router); // calls router
-        testHistory.update("/next", "PUSH");
+        testHistory.navigate("/next", "PUSH");
         expect(testHistory.location).toMatchObject({
           pathname: "/next"
         });
@@ -173,7 +173,7 @@ describe('update(..., "PUSH")', () => {
           pending.finish();
         }
         testHistory.respondWith(router); // calls router
-        testHistory.update("/next", "PUSH");
+        testHistory.navigate("/next", "PUSH");
         expect(testHistory.action).toBe("PUSH");
       });
     });
@@ -187,7 +187,7 @@ describe('update(..., "PUSH")', () => {
           expect(window.location.hash).toBe("#/one");
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two", "PUSH");
+        testHistory.navigate("/two", "PUSH");
       });
 
       it("does not update the action value", () => {
@@ -198,7 +198,7 @@ describe('update(..., "PUSH")', () => {
           expect(testHistory.action).toBe("PUSH");
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two", "PUSH");
+        testHistory.navigate("/two", "PUSH");
       });
     });
   });
@@ -213,7 +213,7 @@ describe('update(..., "PUSH")', () => {
       testHistory.confirmWith(confirm);
       testHistory.respondWith(router); // calls router
 
-      testHistory.update("/next", "PUSH");
+      testHistory.navigate("/next", "PUSH");
       expect(router.mock.calls.length).toBe(2);
     });
 
@@ -226,7 +226,7 @@ describe('update(..., "PUSH")', () => {
       testHistory.confirmWith(confirm);
       testHistory.respondWith(router); // calls router
 
-      testHistory.update("/next", "PUSH");
+      testHistory.navigate("/next", "PUSH");
       expect(router.mock.calls.length).toBe(1);
     });
   });