+++ packages/hash/tests/unit/replace.spec.ts
@@ -15,7 +15,7 @@ function ignoreFirstCall(fn) {
 
 // We create our own jsdom instead of using the one that Jest will create
 // so that we can reset the DOM between tests
-describe('update(..., "REPLACE")', () => {
+describe('navigate(..., "REPLACE")', () => {
   let dom;
   let window;
 
@@ -35,7 +35,7 @@ describe('update(..., "REPLACE")', () => {
   describe("without a response handler", () => {
     it("does nothing", () => {
       const testHistory = Hash();
-      testHistory.update("/two", "REPLACE");
+      testHistory.navigate("/two", "REPLACE");
       expect(window.location.hash).toBe("#/one");
     });
   });
@@ -45,7 +45,7 @@ describe('update(..., "REPLACE")', () => {
       const testHistory = Hash();
       const router = jest.fn();
       testHistory.respondWith(router); // calls router
-      testHistory.update("/two", "REPLACE");
+      testHistory.navigate("/two", "REPLACE");
       expect(router.mock.calls.length).toBe(2);
     });
 
@@ -59,7 +59,7 @@ describe('update(..., "REPLACE")', () => {
           });
         });
         testHistory.respondWith(router);
-        testHistory.update("/two?test=ing", "REPLACE");
+        testHistory.navigate("/two?test=ing", "REPLACE");
       });
 
       it("is a location object created from pushed object", () => {
@@ -71,7 +71,7 @@ describe('update(..., "REPLACE")', () => {
           });
         });
         testHistory.respondWith(router);
-        testHistory.update({ pathname: "/two", hash: "test" }, "REPLACE");
+        testHistory.navigate({ pathname: "/two", hash: "test" }, "REPLACE");
       });
 
       it("key maintains current location's major value, increments the minor value", () => {
@@ -85,7 +85,7 @@ describe('update(..., "REPLACE")', () => {
         const initMajorNum = parseInt(initMajor, 10);
         const initMinorNum = parseInt(initMinor, 10);
 
-        testHistory.update("/next", "REPLACE");
+        testHistory.navigate("/next", "REPLACE");
 
         const current = testHistory.location;
         const [currentMajor, currentMinor] = current.key.split(".");
@@ -105,7 +105,7 @@ describe('update(..., "REPLACE")', () => {
           pending.finish();
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/next", "REPLACE");
+        testHistory.navigate("/next", "REPLACE");
       });
     });
 
@@ -118,15 +118,15 @@ describe('update(..., "REPLACE")', () => {
           expect(window.location.hash).toBe("#/two");
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two", "REPLACE");
+        testHistory.navigate("/two", "REPLACE");
       });
 
-      it("does nothing if update is not called", () => {
+      it("does nothing if navigate is not called", () => {
         const testHistory = Hash();
         let call = 0;
         function router(pending) {}
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two", "REPLACE");
+        testHistory.navigate("/two", "REPLACE");
         expect(window.location.hash).toBe("#/one");
       });
 
@@ -136,7 +136,7 @@ describe('update(..., "REPLACE")', () => {
           pending.finish();
         }
         testHistory.respondWith(router); // calls router
-        testHistory.update("/next", "REPLACE");
+        testHistory.navigate("/next", "REPLACE");
         expect(testHistory.location).toMatchObject({
           pathname: "/next"
         });
@@ -148,13 +148,13 @@ describe('update(..., "REPLACE")', () => {
           pending.finish();
         }
         testHistory.respondWith(router); // calls router
-        testHistory.update("/next", "REPLACE");
+        testHistory.navigate("/next", "REPLACE");
         expect(testHistory.action).toBe("REPLACE");
       });
     });
 
     describe("cancel", () => {
-      it("does not update Hash history when cancel function is called", () => {
+      it("does not navigate Hash history when cancel function is called", () => {
         const testHistory = Hash();
         let router = ignoreFirstCall(function(pending) {
           expect(window.location.hash).toBe("#/one");
@@ -162,10 +162,10 @@ describe('update(..., "REPLACE")', () => {
           expect(window.location.hash).toBe("#/one");
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two", "REPLACE");
+        testHistory.navigate("/two", "REPLACE");
       });
 
-      it("does not update the action value", () => {
+      it("does not navigate the action value", () => {
         const testHistory = Hash();
         let router = ignoreFirstCall(function(pending) {
           expect(testHistory.action).toBe("PUSH");
@@ -173,7 +173,7 @@ describe('update(..., "REPLACE")', () => {
           expect(testHistory.action).toBe("PUSH");
         });
         testHistory.respondWith(router); // calls router
-        testHistory.update("/two", "REPLACE");
+        testHistory.navigate("/two", "REPLACE");
       });
     });
   });
@@ -188,7 +188,7 @@ describe('update(..., "REPLACE")', () => {
       testHistory.confirmWith(confirm);
       testHistory.respondWith(router); // calls router
 
-      testHistory.update("/next", "REPLACE");
+      testHistory.navigate("/next", "REPLACE");
       expect(router.mock.calls.length).toBe(2);
     });
 
@@ -201,7 +201,7 @@ describe('update(..., "REPLACE")', () => {
       testHistory.confirmWith(confirm);
       testHistory.respondWith(router); // calls router
 
-      testHistory.update("/next", "REPLACE");
+      testHistory.navigate("/next", "REPLACE");
       expect(router.mock.calls.length).toBe(1);
     });
   });