+++ packages/in-memory/tests/go.spec.ts
@@ -76,11 +76,11 @@ describe("go", () => {
         done();
       }
       testHistory.respondWith(router);
-      testHistory.update("/two", "PUSH"); // 1.0
-      testHistory.update("/three", "PUSH"); // 2.0
-      testHistory.update("/four", "PUSH"); // 3.0
-      testHistory.update("/five", "PUSH"); // 4.0
-      testHistory.update("/six", "PUSH"); // 5.0
+      testHistory.navigate("/two", "PUSH"); // 1.0
+      testHistory.navigate("/three", "PUSH"); // 2.0
+      testHistory.navigate("/four", "PUSH"); // 3.0
+      testHistory.navigate("/five", "PUSH"); // 4.0
+      testHistory.navigate("/six", "PUSH"); // 5.0
       setup = true;
       testHistory.go(-2);
     });
@@ -110,18 +110,18 @@ describe("go", () => {
           cancelGo = pending.cancel;
           // trigger a push call and don't resolve the go
           testHistory.respondWith(pushRouter);
-          testHistory.update("/seven", "PUSH");
+          testHistory.navigate("/seven", "PUSH");
         });
         const pushRouter = ignoreFirstCall(function(pending) {
           cancelGo("PUSH");
         });
 
         testHistory.respondWith(initialRouter);
-        testHistory.update("/two", "PUSH"); // 1.0
-        testHistory.update("/three", "PUSH"); // 2.0
-        testHistory.update("/four", "PUSH"); // 3.0
-        testHistory.update("/five", "PUSH"); // 4.0
-        testHistory.update("/six", "PUSH"); // 5.0
+        testHistory.navigate("/two", "PUSH"); // 1.0
+        testHistory.navigate("/three", "PUSH"); // 2.0
+        testHistory.navigate("/four", "PUSH"); // 3.0
+        testHistory.navigate("/five", "PUSH"); // 4.0
+        testHistory.navigate("/six", "PUSH"); // 5.0
 
         testHistory.respondWith(goRouter);
         testHistory.go(-2);
@@ -152,8 +152,8 @@ describe("go", () => {
       testHistory.confirmWith(confirm);
       testHistory.respondWith(router);
 
-      testHistory.update("/two", "PUSH"); // 1.0
-      testHistory.update("/three", "PUSH"); // 2.0
+      testHistory.navigate("/two", "PUSH"); // 1.0
+      testHistory.navigate("/three", "PUSH"); // 2.0
       setup = true;
       testHistory.go(-2);
     });
@@ -168,8 +168,8 @@ describe("go", () => {
       }
       testHistory.respondWith(router);
 
-      testHistory.update("/two", "PUSH"); // 1.0
-      testHistory.update("/three", "PUSH"); // 2.0
+      testHistory.navigate("/two", "PUSH"); // 1.0
+      testHistory.navigate("/three", "PUSH"); // 2.0
       // don't add function until we have setup the history
       testHistory.confirmWith(confirm);
       testHistory.go(-2);