+++ packages/hash/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+* `history.go()` and `history.go(0)` now reload the page.
+* Unify `navigate()`, `push()`, and `replace()` under `navigate()`. The optional second argument controls which type of navigation: `ANCHOR` (default), `PUSH`, or `REPLACE`.
+
 ## 1.0.0-beta.6
 
 * Re-export `LocationDetails`.